refactor: isolate public API in mpq.c
Move reader implementation into a dedicated reader module and expose
writer internals through writer.h. Keep only public LIBMPQ_API definitions
in mpq.c while preserving the existing ABI and hiding cross-file helpers.
