+++ src/Makefile.am
@@ -1,6 +1,6 @@
 # Build the shared library and keep private headers out of installations.
 lib_LTLIBRARIES			= libmpq.la
-noinst_HEADERS			= common.h crypt_buf.h endian.h pkware.h extract.h huffman.h mpq-internal.h pack_begin.h pack_end.h platform.h wave.h
+noinst_HEADERS			= common.h crypt_buf.h endian.h pkware.h extract.h huffman.h mpq-internal.h reader.h writer.h pack_begin.h pack_end.h platform.h wave.h
 
 AM_CPPFLAGS			= -I$(top_builddir) -I$(top_srcdir)/include
 
@@ -14,5 +14,6 @@ GENERAL_SRCS =			\
 	extract.c		\
 	pkware.c		\
 		mpq.c			\
+		reader.c		\
 		wave.c			\
 		writer.c
